# Improve performance of ProfileInfo API to load large team config files