--- a/packages/config/src/api/ConfigProfiles.ts
+++ b/packages/config/src/api/ConfigProfiles.ts
@@ -55,7 +55,7 @@ export class ConfigProfiles extends ConfigApi {
     public get(path: string, mustExist?: boolean): { [key: string]: string } {
         if (mustExist !== false && !this.exists(path))
             return null;
-        return this.buildProfile(path, JSONC.parse(JSONC.stringify(this.mConfig.properties.profiles, null, ConfigConstants.INDENT)));
+        return JSONC.parse(JSONC.stringify(this.buildProfile(path, this.mConfig.mProperties.profiles), null, ConfigConstants.INDENT));
     }
 
     // _______________________________________________________________________
@@ -67,7 +67,7 @@ export class ConfigProfiles extends ConfigApi {
      * @returns True if a profile exists. False otherwise.
      */
     public exists(path: string): boolean {
-        return (this.findProfile(path, this.mConfig.properties.profiles) != null);
+        return (this.findProfile(path, this.mConfig.mProperties.profiles) != null);
     }
 
     // _______________________________________________________________________
@@ -93,7 +93,7 @@ export class ConfigProfiles extends ConfigApi {
      *          for example {"host": "lpar.your.domain.net", port: 1234}
      */
     public defaultGet(profileType: string): { [key: string]: string } {
-        const dflt = this.mConfig.properties.defaults[profileType];
+        const dflt = this.mConfig.mProperties.defaults[profileType];
         return dflt != null ? this.get(dflt) : null;
     }