… packages/config/__tests__/__resources__/ProfInfoApp_large_team_config_proj/ProfInfoApp.config.json